fix narrowin conversion on inject

+++ inc/inject.h
@@ -4,7 +4,7 @@
 #include <elf.h>
 
 int inject_section(cliArgs *args);
-void modify_section_header(Elf64_Addr addr, sectionHeader *section, int offset);
+void modify_section_header(Elf64_Addr addr, sectionHeader *section, long offset);
 char *get_section_name(int index);
 sectionHeader *get_section_by_name(char *section_name);
 void set_section_name(sectionHeader *section, char *name);
